stock-forensic-analysis/requirements.txt
# Core dependencies
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Report generation
jinja2>=3.1.2

# Data analysis and visualization
scipy>=1.11.0
matplotlib>=3.7.0

# API and web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
